+++ .github/workflows/sync-remote-dev.yaml
@@ -46,7 +46,7 @@ jobs:
                 run: ssh -p 2222 root@preview.zenmo.com "cd /root/zenmo-website && java -jar /root/remotedev-all.jar"
             -
                 name: Restart backend
-                run: ssh -p 2222 root@preview.zenmo.com "systemctl restart site-server"
+                run: ssh -p 2222 root@preview.zenmo.com "systemctl restart site-backend"
             -
                 # Restarting the frontend is necessary to pick up changes in the Gradle configuration
                 name: Restart frontend
