+++ backend/src/main/resources/.gitignore
@@ -1,0 +1,3 @@
+# JavaScript files are put here by the build process
+*
+!.gitignore
