--- server/src/main/resources/.gitignore
@@ -1,3 +1,0 @@
-# JavaScript files are put here by the build process
-*
-.gitignore
